+++ CHANGELOG.md
@@ -10,6 +10,8 @@ errors in it, without knowing where they were actually assigned.
 works now too.
 * Doc: Documented closures briefly, reference data types and how to do
 object oriented programming.
+* Doc: Documented more reference handling.
+* Feature: Added `std:write\_str` and added std:to_drop string representation.
 
 0.3.4 (2019-11-20)
 ==================